* Adjustment: Tweak the include paths for platformPOSIX.h.
