+++ Engine/source/platformPOSIX/POSIXStrings.cpp
@@ -20,7 +20,7 @@
 // IN THE SOFTWARE.
 //-----------------------------------------------------------------------------
 
-#include "platformX86UNIX/platformX86UNIX.h"
+#include "platformPOSIX/platformPOSIX.h"
 #include "core/strings/stringFunctions.h"
 #include <stdarg.h>
 #include <ctype.h>
