+++ Engine/source/platformPOSIX/POSIXUtils.cpp
@@ -31,7 +31,7 @@
 #include <sys/wait.h>
 #include <stdlib.h>
 
-#include "platformX86UNIX/platformX86UNIX.h"
+#include "platformPOSIX/platformPOSIX.h"
 #include "platformPOSIX/POSIXUtils.h"
 
 UnixUtils *UUtils = NULL;
